pyproject.toml
[tool.poetry]
name = "callofdragonsocr"
version = "0.1.0"
description = ""
authors = ["Deko <gadse.deko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psutil = "^5.9.7"
numpy = "^1.26.2"
pytesseract = "^0.3.10"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
force_single_line = true
force_sort_within_sections = true
lexicographical = true
single_line_exclusions = ["typing"]
order_by_type = false
group_by_package = true
